requirements.txt
streamlit==1.28.2
Pyrebase4==4.7.1
llama_index==0.8.36
openai==0.28.1
PyPDF2==3.0.1
python-dotenv==1.0.0
python-docx==0.8.11
beautifulsoup4==4.12.2
Markdown==3.5
streamlit_lottie==0.0.5 
requests==2.29.0
langchain==0.0.340


